+++ engine/compmgr/src/cengine.js
@@ -211,6 +211,8 @@ module.exports = class {
                         rtn["method"] = module_key;
                         rtn["strict"] = false;
                         rtn["compname"] = compname;
+                        rtn["worker"] = value;
+                        rtn["abspath"] = modpath;
                         for (let [elname, [element]] of Object.entries(
                           regulation
                         )) {